북마크 삭제 수정

+++ src/main/java/com/newSummary/service/BookmarkService.java
@@ -6,8 +6,11 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.newSummary.domain.dto.board.BoardRequestDTO;
 import com.newSummary.domain.dto.bookmark.BookmarkRequestDTO;
 import com.newSummary.domain.dto.bookmark.BookmarkResponseDTO;
+import com.newSummary.domain.dto.bookmark.BookmarkSuccessDTO;
+import com.newSummary.domain.entity.Board;
 import com.newSummary.domain.entity.Bookmark;
 import com.newSummary.domain.entity.User;
 import com.newSummary.repository.BookmarkRepository;
@@ -17,6 +20,7 @@ import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
+
 @RequiredArgsConstructor
 @Service
 public class BookmarkService {
@@ -25,12 +29,12 @@ public class BookmarkService {
 	private final BookmarkRepository bookmarkRepository;
 	private final UserRepository userRepository;
 	
-	// 북마크 생성
 	@Transactional
+	// 북마크 생성
 	public BookmarkResponseDTO createBookmark(BookmarkRequestDTO bookmarkRequestDTO) {
         // User 엔티티를 가져오는 코드
         User user = userRepository.findByUserEmail(bookmarkRequestDTO.getUserEmail())
-                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+                .orElseThrow(() -> new EntityNotFoundException("유저가 존재하지 않습니다"));
         // Bookmark 엔티티를 생성할 때 DTO의 userEmail을 이용
         Bookmark bookmark = Bookmark.builder()
                 .newsObjectId(bookmarkRequestDTO.getNewsObjectId())
@@ -39,12 +43,11 @@ public class BookmarkService {
         bookmarkRepository.save(bookmark);
 		return new BookmarkResponseDTO(bookmark,bookmarkRequestDTO.getUserEmail());
 	}
-	
 	// 회원 북마크 데이터
 	public List<BookmarkResponseDTO> getBookmark(String userEmail) {
         // User 엔티티를 가져오는 코드
         User user = userRepository.findByUserEmail(userEmail)
-                .orElseThrow(() -> new EntityNotFoundException("User not found"));
+                .orElseThrow(() -> new EntityNotFoundException("유저가 존재하지 않습니다"));
         List<Bookmark> bookmarklist = bookmarkRepository.findByUser(user);
         return convertToBookmarkResponseDTOList(bookmarklist);
 	}
@@ -55,15 +58,15 @@ public class BookmarkService {
                 .collect(Collectors.toList());
     }
     // 북마크 삭제
-	public void deleteBookmark(Long bookmark_idx) {
+	public BookmarkSuccessDTO deleteBookmark(Long bookmark_idx, String userEmail) {
+		Bookmark bookmark = bookmarkRepository.findById(bookmark_idx).orElseThrow(
+				() -> new IllegalArgumentException("북마크가 존재하지 않습니다.")
+		);
+        if (!bookmark.getUser().getUserEmail().equals(userEmail)) {
+            throw new IllegalArgumentException("이 북마크를 삭제할 권한이 없습니다.");
+        }
 		bookmarkRepository.deleteById(bookmark_idx);
-	}
-	// 북마크id를 이용하여 userEmail 가져오기
-	public String getUserEmailByBookmarkId(Long bookmark_idx) {
-        Bookmark bookmark = bookmarkRepository.findById(bookmark_idx)
-                .orElseThrow(() -> new EntityNotFoundException("Bookmark not found"));
-
-        return bookmark.getUser().getUserEmail();
+		return new BookmarkSuccessDTO(true);
 	}
 
 }
